+++ StickManWars/src/stickmanwars/Castle.java
@@ -1,7 +1,6 @@
 package stickmanwars;
 
-import java.awt.Graphics;
-import java.awt.Rectangle;
+import java.awt.*;
 
 public final class Castle extends GameObject{
 
@@ -25,7 +24,7 @@ public final class Castle extends GameObject{
 
     @Override
     public Rectangle getBounds() {
-        return new Rectangle(this.getX(), this.getY(), this.getDimention().width, this.getDimention().height);
+        return new Rectangle((int) this.getX(), (int) this.getY(), this.getDimention().width, this.getDimention().height);
     }
 
     public int getCastleHealth() {
