+++ StickManWars/src/stickmanwars/Weapon.java
@@ -1,8 +1,7 @@
 
 package stickmanwars;
 
-import java.awt.Graphics;
-import java.awt.Rectangle;
+import java.awt.*;
 
 
 public class Weapon extends GameObject{
@@ -29,7 +28,7 @@ public class Weapon extends GameObject{
 
     @Override
     public Rectangle getBounds() {
-        return new Rectangle(this.getX(), this.getY(), this.getDimention().width, this.getDimention().height);
+        return new Rectangle((int) this.getX(), (int) this.getY(), this.getDimention().width, this.getDimention().height);
     }
 
     public int getDamage() {
